[Phase 0.5 / 0.5.4f] MobileMenuButton: rename .tsx + DS Icon + framer drop
- git rm MobileMenuButton.jsx, add MobileMenuButton.tsx
- Drop framer-motion (motion.button, motion.div, iconVariants) — CSS-only toggle
- Drop @/deprecated-components/ui/{BurgerMenuIcon,CancelIcon} — DS <Icon> from
  @umichkisa-ds/web
- Animate open↔close via a pair of absolute-positioned icons cross-fading on
  transition-opacity duration-300
- Typed props explicitly; add aria-label + aria-expanded for accessibility
- Native <button type="button">, same onClick contract; Header.jsx consumer
  shape unchanged (isMobileMenuOpen, setIsMobileMenuOpen)

Refs: issue #53
